[hotfix] Make TestingDispatcher and TestingJobManagerRunnerFactory standalone testing classes
Refactors the DispatcherTests and moves the TestingDispatcher and the TestingJobManagerRunnerFactory
to be top level classes. This makes it easier to reuse them.
